refactor table row component, heart feature
