+++ src/components/pages/order/index.ts
@@ -3,4 +3,7 @@ export * from './menu-dropdown';
 export * from './shared-link';
 export * from './shop-owner';
 export * from './table';
+export * from './table-add-row-button';
+export * from './table-header';
+export * from './table-row';
 export * from './transfer-info';
